Move to_gaussian_similarity into linfa-nn and parametrize it by distance
