+++ algorithms/linfa-reduction/src/lib.rs
@@ -11,4 +11,3 @@ pub mod utils;
 pub use diffusion_map::{DiffusionMap, DiffusionMapParams, DiffusionMapValidParams};
 pub use error::{ReductionError, Result};
 pub use pca::{Pca, PcaParams};
-pub use utils::to_gaussian_similarity;
